added the possibilities to edit and delete a user

+++ src/main/java/com/example/onlineshop/repository/UserRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
-User findByLogin(String login);
+public interface UserRepository extends JpaRepository<User, Long> {
+    User findByLogin(String login);
 }
+
